requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
concurrent-log-handler==0.9.9
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
elasticsearch==7.13.4
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
kafka-python==1.4.6
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pika==1.2.0
psycopg2==2.9.1
py-log==2.5
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.5.1
python-decouple==3.4
pytz==2021.1
regex==2021.7.6
requests==2.26.0
sqlparse==0.4.1
tomli==1.2.0
tomorrow3==1.1.0
urllib3==1.26.6
